==> 1_etl_movies.ipynb <==
Importamos las librerias que vamos a utilizar.

import pandas as pd
import numpy as np
import ast


# Establecer la opción para mostrar todas las columnas
pd.set_option('display.max_columns', None)

# Carga de datos.

movies_raw = pd.read_csv('data/movies_dataset.csv',low_memory=False)

---

# Analisis de movies_raw.

movies_raw.info()

Tenemos 45466 filas y 24 columnas. 
Algunas columnas presentan inconsistencias en sus datos y valores nulos.

movies_raw.describe(include='all')

movies_raw.head()

Podemos observar 5 columnas que contienen datos anidados. Estas son: 'belongs_to_collection', 'genres', 'production_companies', 'production_countries' y 'spoken_languages'.

---

# Valores Duplicados y nulos.

¿Hay filas duplicadas? ¿Cuántas?


#filas duplicadas
(
    movies_raw
    .duplicated(keep=False)
    .sum()
)

# Filas con campo id duplicado /  Una misma pelicula aparece varias veces en el dataframe.
(
    movies_raw
    .duplicated(subset='id',keep=False)
    .sum()
)

Eliminaremos las filas duplicadas y las que poseen ids duplicados. 
Ya que representan menos del 1% del dataset.
Pero se consultará para posteriores entregas.

¿Cuántos datos nulos tiene cada columna?

(
    movies_raw
    .isnull()
    .sum()
)

¿Cuáles son las 6 peliculas que no tienen título especificado?

movies_raw[movies_raw['title'].isnull()]

Estas 6 filas también se eliminaran ya que la mayoría de sus datos son nulos o inconsistentes.

---

# Creación del dataframe movies.

movies = (
    movies_raw[
        ['id',
        'title',
        'tagline',
        'overview',
        'original_language',
        'runtime',
        'status',
        'release_date',
        # release_year
        # id_collection
        # return
        'budget',
        'revenue',
        'vote_count',
        'vote_average', 
        'popularity',
        'belongs_to_collection',
        'genres',
        'spoken_languages',
        'production_companies',
        'production_countries'
        ]]
)



movies.shape

# Datos duplicados y nulos

Eliminamos las filas duplicadas y las que poseen ids duplicados ya que representan menos del 1% de los datos. 
Pero se tendran en cuenta para próximas entregas.

movies = movies.drop_duplicates()

movies = movies.drop_duplicates(subset='id',keep=False)

# movies[movies['id']==str(10991)]  # revisamos si se borraron todas las filas duplicadas con un id de prueba

# # revimamos si quedaron ids duplicados 
# print(movies[
#     movies
#     .duplicated(subset='id',keep=False)   
# ].shape)
# # ids_duplicados.sort_values(by='id')[['id','title']] #para ordenarlos

movies.shape

Eliminamos las 6 filas que poseen valores nulos en el campo title. Ya que poseen valores inconsistentes y muchos valores nulos.

movies = movies.drop(movies[movies['title'].isnull()].index)
movies.shape

El campo release_date tiene 87 valores nulos que deben eliminarse
Los valores nulos del campo release date deben eliminarse. Debido a que es item de la Propuesta de trabajo.

# movies[movies['release_date'].isnull()].head(5) # filtrar las filas con release_date NAN

#Eliminar los release_date NAN // release_date      
movies = movies.dropna(subset='release_date')

movies.info()

# Tipo de datos y creacion de la columna 'release_year'

Cambiamos los tipos de los datos para que estos sean consistentes con los datos que contienen.

#cambiamos los tipos de datos de las columnas
movies = movies.rename({'id': 'id_movie'}, axis=1)

movies['title'] = movies['title'].apply(lambda x: x.title())

movies = movies.astype({'id_movie' : 'int64', 
                        # 'title':'str',
                        # 'tagline':'str',
                        # 'overview':'str',
                        'original_language':'category',
                        # # 'runtime':'int', # posee valores nan por lo que aun no podemos cambiarlo a int
                        'status': 'category', 
                        # 'release_date': ,  # no se si se puede hacer aca directamente
                        'budget': 'int64', 
                        'revenue': 'int64', # posee valores nan por lo que aun no podemos cambiarlo a int
                        'vote_count': 'int64', # posee valores nan por lo que aun no podemos cambiarlo a int
                        'vote_average': 'int', # posee valores nan por lo que aun no podemos cambiarlo a int
                        # 'popularity' : 'float' # no se que es esto jajaj 
                        }, copy=False, errors='raise')

De haber fechas, deberán tener el formato AAAA-mm-dd.

#cambiamos los tipos de datos de la columna 'release_date'
movies['release_date'] = pd.to_datetime(movies['release_date'], format='%Y-%m-%d',yearfirst=True, errors='coerce')

A demás deberán crear la columna release_year donde extraerán el año de la fecha de estreno

#Creamos la columna 'release_year'
movies.insert(movies.columns.get_loc('release_date') + 1, 'release_year', movies['release_date'].dt.year)

movies.head(3)

---

Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0.

# Los valores nulos de los campos revenue, budget deben ser rellenados por el número 0
movies['budget'].fillna(0, inplace=True)
movies['revenue'].fillna(0, inplace=True)


---


Crear la columna con el retorno de inversión, llamada return con los campos revenue y budget, dividiendo estas dos últimas revenue / budget, cuando no hay datos disponibles para calcularlo, deberá tomar el valor 0.

# funcion que permite calcular el return
def ret(df,col1,col2):
    if df[col1] == 0:
        out = 0

        
    else:    
        out = df[col2] / df[col1]
    return out

# creamos la columna return
movies.insert(movies.columns.get_loc('revenue') + 1, 'return', movies.apply(ret,axis=1,args={'budget','revenue'}))

----

# Creación de tablas a partir de las columnas con datos anidados

----------

# Creación de la tabla 'release_date_table'

#TABLA LANZAMIENTOS

release_date_table = movies[['id_movie','release_date']].copy()

release_date_table.loc['year'] = movies['release_date'].dt.year

release_date_table['month'] = movies['release_date'].dt.month_name(locale='es_ES.UTF-8').str.lower()

release_date_table['day_num'] = movies['release_date'].dt.day

release_date_table['day_name'] = movies['release_date'].dt.day_name(locale='es_ES.UTF-8').str.lower()
release_date_table['day_name'] = release_date_table['day_name'].replace({'miã©rcoles': 'miércoles', 'sã¡bado': 'sábado'})


release_date_table

(
    release_date_table
    .isnull()
    .sum()
)

release_date_table = release_date_table.dropna()

#crear el csv 
release_date_table.to_csv("data/release_date.csv",index=False)

Creacion de las tablas estrenos_por_dia y estrenos_por_mes que se utilizaran para las consultas.

estrenos_por_dia = release_date_table['day_name'].value_counts().reset_index()
estrenos_por_dia.columns = ['dia', 'cantidad']
estrenos_por_dia

estrenos_por_mes = release_date_table['month'].value_counts().reset_index()
estrenos_por_mes.columns = ['mes', 'cantidad']
estrenos_por_mes

#crear los csv 
estrenos_por_dia.to_csv("data/estrenos_por_dia.csv",index=False)
estrenos_por_mes.to_csv("data/estrenos_por_mes.csv",index=False)


---

# Creación de la tabla 'collections'

# TABLA COLLECTIONS

# Extrae los datos anidados en la columna "belongs_to_collection"
collection = pd.json_normalize(data=movies['belongs_to_collection'].dropna().apply(eval))

collection = pd.DataFrame(collection)

collection = collection[['id','name']]
collection = collection.rename({'id':'id_collections'},axis=1)

#crear el csv 
collection.to_csv("data/collection.csv",index=False)

# belongs_to_collection.head()
# belongs_to_collection.shape

# cremaos la columna 'id_collecction', extraemos sus valores de la columna belongs_to_collection y luego eliminamos esta última.

# # Definimos una expresión regular para extraer el id del campo "id" de la columna belongs_to_collection
regex = r"'id':\s*([\d]+)"
'''
r": esto indica que la cadena debe ser interpretada como una "cadena sin procesar" en Python, lo que significa que cualquier carácter especial se interpreta literalmente, sin ser escapado.
'id': esto es una cadena literal que busca el texto exacto 'id'.
:\s*: esto busca un carácter de dos puntos seguido de cero o más espacios en blanco. La secuencia \s* busca cualquier número de caracteres de espacio en blanco (espacios, tabuladores, saltos de línea, etc.) que puedan aparecer después del carácter de dos puntos.
([\d]+): esto busca uno o más caracteres numéricos, que representan el valor de id. La expresión [\d]+ indica que se busca uno o más caracteres que sean dígitos numéricos (\d es una abreviatura para "dígitos numéricos").
En resumen, la expresión regular busca la cadena 'id': seguida de cualquier número de espacios en blanco, y luego busca uno o más dígitos numéricos para capturar el valor de id. Cuando se aplica esta expresión regular a una cadena JSON que contiene un valor de id, se captura y extrae el valor de id.
'''

# # Utilizamos la función "str.extract" para extraer el id y guardarlo en una nueva columna 'id_collection'
movies['id_collection'] = movies['belongs_to_collection'].str.extract(regex)

# # moverla al principio para comparar
# movies = movies.insert(0, 'id_collection', movies.pop('id_collection'))

# Eliminamos la columna 'belongs_to_collection'
movies.drop(['belongs_to_collection'], axis= 'columns', inplace=True)

# movies.head(2)

--------

# Creación de la tabla 'movies_production_companies'

### TABLA MOVIES_production_companies
movies_production_companies = movies[['id_movie','production_companies']].copy()

#reemplazamos los valores nulos por "[{'name': 'None', 'id': 0}]"
movies_production_companies['production_companies'] = movies_production_companies[['production_companies']].replace("[]", "[{'name': 'None', 'id': 0}]")

movies_production_companies['production_companies'] = movies_production_companies['production_companies'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista production_companies
movies_production_companies = movies_production_companies.explode('production_companies')

# Extraemos el valor de id_production_companies de cada elemento en la lista production_companies
movies_production_companies['id_production_companies'] = movies_production_companies['production_companies'].apply(lambda x: x['id'])

# # Eliminar la columna production_companies
movies_production_companies = movies_production_companies.drop('production_companies', axis=1)

# Creamos el csv
movies_production_companies.to_csv("data/movies_production_companies.csv",index=False)


# TABLA production_companies

production_companies = movies['production_companies'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
production_companies = pd.DataFrame([dict(x) for x in production_companies.sum()])[['id', 'name']]
production_companies = production_companies.drop_duplicates()
production_companies.to_csv("data/production_companies.csv",index=False)


# Eliminamos la columna production_companies de la tabla movies
movies = movies.drop('production_companies', axis=1)

-------

# Creación de la tabla 'movies_production_countries'

### TABLA MOVIES_production_countries

movies_production_countries = movies[['id_movie','production_countries']].copy()

# Reemplazamos los valores nulos por "[{'iso_3166_1': 'none', 'name': 'none'}]"
movies_production_countries['production_countries'] = movies_production_countries[['production_countries']].replace("[]", "[{'iso_3166_1': 'none', 'name': 'none'}]")  # ya esta arriba
movies_production_countries['production_countries'] = movies_production_countries['production_countries'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista production_countries
movies_production_countries = movies_production_countries.explode('production_countries')

# Extraer el valor de id_production_countries de cada elemento en la lista production_countries
movies_production_countries['id_production_countries'] = movies_production_countries['production_countries'].apply(lambda x: x['iso_3166_1'])

# Eliminar la columna genres
movies_production_countries = movies_production_countries.drop('production_countries', axis=1)

# Creamos el csv
movies_production_countries.to_csv("data/movies_production_countries.csv",index=False)


### TABLA production_countries

production_countries = movies['production_countries'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
production_countries = pd.DataFrame([dict(x) for x in production_countries.sum()])[['iso_3166_1', 'name']]
production_countries = production_countries.drop_duplicates()

# Creamos el csv
production_countries.to_csv("data/production_countries.csv",index=False)

# Eliminamos la columna production_countries de la tabla movies
movies = movies.drop('production_countries', axis=1)


-------

# Creación de la tabla 'movies_spoken_languages'

### TABLA MOVIES_spoken_languages

movies_spoken_languages = movies[['id_movie','spoken_languages']].copy()

# Reemplazamos los valores nulos por "[{'iso_639_1': 'none', 'name': 'none'}]"
movies_spoken_languages['spoken_languages'] = movies_spoken_languages[['spoken_languages']].replace("[]", "[{'iso_639_1': 'none', 'name': 'none'}]")  
movies_spoken_languages['spoken_languages'] = movies_spoken_languages['spoken_languages'].apply(ast.literal_eval)

# Aplicamos explode() para crear filas adicionales para cada elemento en la lista genres
movies_spoken_languages = movies_spoken_languages.explode('spoken_languages')

# Extraer el valor de id_genre de cada elemento en la lista genres
movies_spoken_languages['id_spoken_languages'] = movies_spoken_languages['spoken_languages'].apply(lambda x: x['iso_639_1'])

# Eliminar la columna spoken_languages
movies_spoken_languages = movies_spoken_languages.drop('spoken_languages', axis=1)

# Creamos el csv
movies_spoken_languages.to_csv("data/movies_spoken_languages.csv",index=False)


### TABLA spoken_languages

spoken_languages = movies['spoken_languages'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
spoken_languages = pd.DataFrame([dict(x) for x in spoken_languages.sum()])[['iso_639_1', 'name']]
spoken_languages = spoken_languages.drop_duplicates()

# Creamos el csv
spoken_languages.to_csv("data/spoken_languages.csv",index=False)


# Eliminamos la columna production_countries de la tabla movies
movies = movies.drop('spoken_languages', axis=1)

-------

# Creación de la tabla 'genres'

# TABLA GENRES

genres = movies['genres'].apply(ast.literal_eval) # Convertir el string a lista de diccionarios
genres = pd.DataFrame([dict(x) for x in genres.sum()])[['id', 'name']]
genres = genres.drop_duplicates()
genres.to_csv("data/genres.csv",index=False)


### TABLA MOVIES_GENRES
movies_genres = movies[['id_movie','genres']].copy()

# Generamos la categoria none para las peliculas que no tienen un genero especificado
movies_genres['genres'] = movies_genres[['genres']].replace("[]", "[{'id': 0, 'name': 'None'}]")  # ya esta arriba

movies_genres['genres'] = movies_genres['genres'].apply(ast.literal_eval)

# # Aplicamos explode() para crear filas adicionales para cada elemento en la lista genres
movies_genres = movies_genres.explode('genres')

# # Extraer el valor de id_genre de cada elemento en la lista genres
movies_genres['id_genre'] = movies_genres['genres'].apply(lambda x: x['id'])

# Eliminar la columna genres
movies_genres = movies_genres.drop('genres', axis=1)


# 
movies = movies.drop('genres', axis=1)


movies_genres.to_csv("data/movies_genres.csv",index=False)



movies.head()

movies.to_csv("data/movies_etl.csv",index=False)
# .


movies

--------------
