==> 2_etl_credits.ipynb <==
Importamos las librerias que vamos a utilizar.

import pandas as pd
import numpy as np
import ast

# Cargamos los datasets que vamos a utilizar.

credits_raw = pd.read_csv('data/credits.csv',low_memory=False)

# Analisis de credits_raw

credits_raw.head(2)

credits_raw.info()

Analisis de la columna cast

# cast
print (credits_raw.iloc[0,0]) # ejemplo de los datos
print (type(credits_raw.iloc[0,0])) # tipo de los datos

Analisis de la columna crew

# crew
print (credits_raw.iloc[0,1]) # ejemplo de los datos
print (type(credits_raw.iloc[0,1])) # tipo de los datos


-----------

# Extracción de los datos de la columna cast. Creacion de la tabla actors.

print (ast.literal_eval(credits_raw.iloc[0,0])[0].keys()) # Obtener las llaves del diccionario

credits = credits_raw.rename(columns={'id': 'id_movie'}) #  Cambiamos el nombre de la columna id de credits por id_movies para evitar confusiones.
credits['cast'] = credits['cast'].apply(ast.literal_eval) # Convertimos los str dentro de cast en estructuras de python correspondiente, en este caso en una lista
actors = credits.explode('cast')['cast'].apply(pd.Series)[['id','name']] # Explode('cast') desanida la columna cast y crea un nuevo df con filas duplicadas segun los elementos de la lista en la columna cast. De este nuevo df solo seleccionamos las columnas que vamos a utilizar.
actors['id_movie'] = credits['id_movie'] # Agrega la columna id_movie al df actors
actors = actors.rename(columns={'id': 'id_actor'}) # Cambiamos el nombre de la columna id por id_actor


actors.head(2)

Exploramos la tabla actors

actors.shape

actors.info()

actors.max()

¿Hay valores nulos? ¿Cuántos?

(
    actors
    .isnull()
    .sum()
)

¿Cuáles son las filas con valores nulos?

actors[
    actors
    .isnull()
    .any(axis=1)
]

Eliminamos las filas con valores nulos ya que no aportan información relevante.

actors = actors.dropna(subset='id_actor')

Corregimos los tipo de dato de las columnas

actors = actors.astype({'id_actor':'int64', 'name':'str'}, copy=False, errors='raise')

Existen filas duplicadas?

actors[actors.duplicated()].count()

actors = actors.drop_duplicates()

actors.sort_values('id_actor')

Creamos el actors.csv

actors.to_csv("data/actors.csv",index=False)

-------------

# Extracción de los datos de la columna crew. Creacion de la tabla directors.

print (ast.literal_eval(credits_raw.iloc[0,1])[0].keys()) # Obtener las llaves del diccionario

credits['crew'] = credits['crew'].apply(ast.literal_eval) # Convertimos los str dentro de cast en estructuras de python correspondiente, en este caso en una lista
directors = credits.explode('crew')['crew'].apply(pd.Series)[['id','name']] # Explode('cast') desanida la columna cast y crea un nuevo df con filas duplicadas segun los elementos de la lista en la columna cast. De este nuevo df solo seleccionamos las columnas que vamos a utilizar.
directors['id_movie'] = credits['id_movie'] # Agrega la columna id_movie al df actors
directors = directors.rename(columns={'id': 'id_director'}) # Cambiamos el nombre de la columna id por id_actor

directors.head(2)

Exploramos la tabla directors

directors.info()

directors.max()

¿Hay valores nulos? ¿Cuántos?

(
    directors
    .isnull()
    .sum()
)

Eliminamos las filas con valores nulos ya que no aportan información relevante.

directors = directors.dropna(subset='id_director')

Corregimos los tipo de dato de las columnas

directors = directors.astype({'id_director':'int64', 'name':'str'}, copy=False, errors='raise')

Existen filas duplicadas?

directors[directors.duplicated()].count()

directors = directors.drop_duplicates()

directors.info()

directors.to_csv("data/directors.csv",index=False)

-------
